backend/requirements.txt
aiofiles==24.1.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
Flask==3.1.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
identity==0.9.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msal==1.31.1
oauthlib==3.2.2
priority==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pymongo==4.10.1
pyucc==1.7
Quart==0.19.9
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
siblink @ git+https://github.com/TreltaSev/siblink.git@c90aa3123ddb20a82692aa5d81eb7b49af252484
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.1.3
wsproto==1.2.0
